colocate types with commands
